backend/shop/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-cors-headers==4.7.0
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
mysqlclient==2.2.7
pillow==11.2.1
PyJWT==2.9.0
redis==6.2.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
